refactor(command): remove invalid Command interface and simplify command layer
