--- src/Command/Command.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Haspadar\Piqule\Command;
-
-interface Command
-{
-    public function run(): void;
-}
